+++ include/algorithms/penalizedGraph/penalizedGraphAlgorithm.ipp
